pyproject.toml
[tool.poetry]
name = "mopsimcbot"
version = "0.1.0"
description = ""
authors = ["PederHA <peder.andresen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.3"
async-timeout = "^3.0.1"
pathvalidate = "^2.3.0"
unidecode = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
